refactor(navigation): restructure navigation folder with proper directory organization and type fixes
