--- src/components/Dashboard/Navigation/Navigation.types.ts
@@ -1,7 +1,0 @@
-export type NavigationItem = {
-  id: string;
-  label: string;
-  icon: React.ReactNode;
-  section: "main" | "game" | "social";
-  path?: string;
-};
