+++ src/components/Dashboard/Navigation/components/NavigationButton/index.ts
@@ -1,0 +1,1 @@
+export { NavigationButton } from './NavigationButton.component';
