+++ src/components/Dashboard/Navigation/components/NavigationLogo/index.ts
@@ -1,0 +1,1 @@
+export { NavigationLogo } from './NavigationLogo.component';
