+++ src/components/Dashboard/Navigation/components/NavigationSection/index.ts
@@ -1,0 +1,1 @@
+export { NavigationSection } from './NavigationSection.component';
