+++ src/components/Dashboard/Navigation/constants/navigationSections.ts
@@ -1,0 +1,5 @@
+export const navigationSections = {
+  main: "Główne",
+  game: "Gra",
+  social: "Społeczność",
+} as const;
